pyproject.toml
[project]
name = "x-bot"
version = "0.0.1"
description = "Small 𝕏 bot"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aiosqlite>=0.19.0",
    "apscheduler>=3.11.1",
    "argon2-cffi>=23.1.0",
    "langchain>=0.3.7",
    "langchain-anthropic>=0.3.0",
    "langchain-chroma>=0.1.4",
    "langchain-community>=0.3.7",
    "langchain-google-genai>=2.0.4",
    "langchain-openai>=0.2.5",
    "langgraph>=0.2.25",
    "pydantic-settings>=2.7.1",
    "openai>=2.7.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "selenium>=4.38.0",
    "selenium-stealth>=1.0.6",
    "setuptools>=65.0.0",
    "tenacity>=8.2.0",
    "twscrape>=0.17.0",
    "itsdangerous>=2.2.0",
    "undetected-chromedriver>=3.5.5",
    "webdriver-manager>=4.0.2",
    "fastapi>=0.122.0",
    "uvicorn>=0.38.0",
    "jinja2>=3.1.6",
    "chromadb>=1.3.5",
    "httpx>=0.28.1",
    "python-multipart>=0.0.20",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "httpx>=0.27.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
